changed file ending to jsonchgs

+++ src/test/java/org/sdmlib/test/examples/modelspace/chat/MSChatClientTest.java
@@ -41,7 +41,10 @@ public class MSChatClientTest
       {
          for (File f : file.listFiles())
          {
-            f.delete();
+            if ( ! f.getName().endsWith(".gitignore"))
+            {
+               f.delete();
+            }
          }
       }
